Allow cli use without solc-backend, sans bytecode support
